Put renderer code in separate files
